Had to change Long to Double here too.
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@20493 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/jspwrap/NumberRangeParamBean.java
@@ -25,11 +25,11 @@ public class NumberRangeParamBean extends ParamBean<NumberRangeParam> {
         return _param.isInteger();
     }
     
-    public Long getMin() {
+    public Double getMin() {
       return _param.getMin();
     }
     
-    public Long getMax() {
+    public Double getMax() {
       return _param.getMax();
     }
